rename discarders
