--- internal/pkg/archiver/discard/discarder/README.md
@@ -1,3 +1,0 @@
-# Discarder
-
-To avoid conflicts with the `postprocessor/sitespecific/*` packages, All the custom discarders packages are prefixed with `dc_`
