+++ internal/pkg/archiver/discard/discarder/cloudflare/cloudflare.go
@@ -1,4 +1,4 @@
-package dc_cloudflare
+package cloudflare
 
 import "net/http"
