+++ internal/pkg/archiver/discard/reasoncode/reasoncode.go
@@ -3,7 +3,7 @@ package reasoncode
 import (
 	"slices"
 
-	"github.com/internetarchive/Zeno/internal/pkg/archiver/discard/discarder/dc_cloudflare"
+	"github.com/internetarchive/Zeno/internal/pkg/archiver/discard/discarder/cloudflare"
 )
 
 var HookNotSet = "Hook not set"
@@ -13,7 +13,7 @@ var AllPassed = "All hooks passed, no need to discard"
 // IsChallengePage checks if the response is a challenge page.
 func IsChallengePage(reason string) bool {
 	reasons := []string{
-		dc_cloudflare.ChallengeDetected,
+		cloudflare.ChallengeDetected,
 	}
 	return slices.Contains(reasons, reason)
 }
